Massive Overhaul: Factory-Based Design Objects
Completely refactored the project. All design classes now reside in factories which managed their meta-data and other uses. It's still not as clean as I want it, but it avoids multiple hierarchies, keeps the base design object clean of specializations, finalized and private. It throws the single responsibility principle kinda out the bus, and it still uses too much downcasting, but code resuse is high, so that's great.